+++ yazi-fs/src/file/inventory.rs
@@ -4,7 +4,7 @@ use crate::file::File;
 
 pub struct FileInventory {
 	pub register: fn(&mut UserDataRegistry<File>),
-	pub from_lua: fn(&AnyUserData) -> mlua::Result<File>,
+	pub borrow:   fn(&AnyUserData, &mut dyn FnMut(&File) -> mlua::Result<()>) -> mlua::Result<()>,
 }
 
 inventory::collect!(FileInventory);