+++ yazi-fs/src/file/mod.rs
@@ -1,1 +1,1 @@
-yazi_macro::mod_flat!(data file_cov file files inventory lua);
+yazi_macro::mod_flat!(data file_cov file file_ref files inventory lua);