+++ yazi-plugin/src/theme/mod.rs
@@ -1,1 +1,1 @@
-yazi_macro::mod_flat!(theme);
+yazi_macro::mod_flat!(icon theme);