+++ yazi-plugin/src/theme/theme.rs
@@ -3,7 +3,7 @@ use yazi_binding::{Composer, ComposerGet, ComposerSet, style::Style};
 use yazi_config::{THEME, theme::CustomSectionArc};
 use yazi_shared::url::UrlBuf;
 
-use crate::LUA;
+use crate::{LUA, theme::icon};
 
 pub fn compose() -> Composer<ComposerGet, ComposerSet> {
 	fn get(lua: &Lua, key: &[u8]) -> mlua::Result<Value> {
@@ -23,6 +23,7 @@ pub fn compose() -> Composer<ComposerGet, ComposerSet> {
 			b"cmp" => cmp(),
 			b"tasks" => tasks(),
 			b"help" => help(),
+			b"icon" => icon(),
 			_ => return custom(lua, key),
 		}
 		.into_lua(lua)